hasta ej 4

+++ practico_04/ejercicio_03.py
@@ -2,9 +2,9 @@
 
 import datetime
 
-from practico_04.ejercicio_01 import reset_tabla
-from practico_04.ejercicio_02 import agregar_persona
-
+from ejercicio_01 import reset_tabla
+from ejercicio_02 import agregar_persona
+import sqlite3
 
 def borrar_persona(id_persona):
     """Implementar la funcion borrar_persona, que elimina un registro en la 
@@ -13,10 +13,19 @@ def borrar_persona(id_persona):
     pass # Completar
     conn=sqlite3.connect('basededatos.db')
     cursor=conn.cursor()
-    query=f"DELETE FROM persona WHERE id_persona={id_persona}"
+    query=f"SELECT id_persona FROM persona WHERE id_persona={id_persona}"
     cursor.execute(query)
-    conn.commit()
-    conn.close()
+    if cursor.fetchone() is None:
+        conn.commit()
+        conn.close()
+        return False
+
+    else:
+        query2=f"DELETE FROM persona WHERE id_persona={id_persona}"
+        cursor.execute(query2)
+        conn.commit()
+        conn.close()
+        return True
     
     
 # NO MODIFICAR - INICIO
